+++ test/games/chase.test.ts
@@ -1,4 +1,5 @@
-/* tslint:disable:no-unused-expression */
+/* eslint-disable no-unused-expressions */
+/* eslint-disable @typescript-eslint/no-unused-expressions */
 
 import "mocha";
 import { expect } from "chai";